pyproject.toml
[tool.poetry]
name = "pybotx-widgets"
version = "0.3.1"
description = "Widgets for pybotx"
authors = ["Tamirlan Dzhemirzoev", "Alexandr Samojlenko", "Vladimir Karabanov", "Samat Yunusov"]

[tool.poetry.dependencies]
python = "^3.8"
botx = ">=0.14.1, <0.20.0"
python-dateutil = "^2.8.1"
mako = "^1.1.3"

[tool.poetry.dev-dependencies]
black = "~20.8b1"
isort = "4.3.21"
autoflake = "~1.4"
mypy = "~0.782"
wemake-python-styleguide = "~0.14"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
